add first two pages of spread
