+++ src/App.js
@@ -1,12 +1,14 @@
 import React from 'react';
 import CreatorStore from './Components/CreatorStore';
 import SpreadAcademy from './Components/SpreadAcademy';
+import { GlobalStyles } from './Components/CreatorsStoreStyled';
 
 const App = () => {
   return (
     <div>
-      <CreatorStore />
+      <GlobalStyles />
       <SpreadAcademy />
+      <CreatorStore />
     </div>
   );
 };
