+++ src/font/Proxima Nova/COPYRIGHT.txt
@@ -1,0 +1,1 @@
+Copyright (c) Mark Simonson, 2005. All rights reserved.
